+++ src/container/home-page-container.js
@@ -1,0 +1,10 @@
+/**
+ *
+ * Copyright 2017 Prottonne
+ */
+
+ export const mapStateToProps = (store) => {
+   return {
+     user:store.user
+   };
+ }
