+++ src/container/login-form-container.js
@@ -3,7 +3,6 @@
  * Copyright 2017 Prottonne
  */
 
-import { setUserEmail, setUserPassword } from '../action/user-action';
 import { login } from '../action/auth-action';
 
  export const mapStateToProps = (store) => {
@@ -14,8 +13,6 @@ import { login } from '../action/auth-action';
 
  export const matchDispatchToProps = (dispatch) => {
    return   {
-     setUserEmail: (email) => dispatch(setUserEmail(email)),
-     setUserPassword: (password) => dispatch(setUserPassword(password)),
      login: (user) => dispatch(login(user))
    };
  }
